+++ tilelang/tladapter/conversion.py
@@ -7,6 +7,6 @@ Each pass accepts str and Module. Use pass(**opts) for pipeline with options.
 Add lowering passes: name = pass_fn("pass-name").
 """
 
-from tilelang.tilelangir.utils import pass_fn
+from tilelang.tladapter.utils import pass_fn
 
 # Conversion passes - add lowering passes here
