+++ tilelang/tladapter/transforms/bishengir.py
@@ -2,6 +2,6 @@
 # Licensed under the MIT License.
 """BishengIR dialect transformation passes."""
 
-from tilelang.tilelangir.utils import pass_fn
+from tilelang.tladapter.utils import pass_fn
 
 adapt_triton_kernel = pass_fn("adapt-triton-kernel")
