+++ tilelang/tladapter/transforms/tilelangir.py
@@ -2,7 +2,7 @@
 # Licensed under the MIT License.
 """TileLangIR dialect transformation passes."""
 
-from tilelang.tilelangir.utils import pass_fn
+from tilelang.tladapter.utils import pass_fn
 
 cv_split = pass_fn("tilelangir-cv-split")
 vectorize = pass_fn("tilelangir-vectorize")
